+++ alibi_detect/cd/pytorch/lsdd_online.py
@@ -2,11 +2,11 @@ from tqdm import tqdm
 import numpy as np
 import torch
 from typing import Any, Callable, Optional, Union
-from alibi_detect.cd.base_online import BaseDriftOnline
+from alibi_detect.cd.base_online import BaseMultiDriftOnline
 from alibi_detect.utils.pytorch import GaussianRBF, permed_lsdds, quantile
 
 
-class LSDDDriftOnlineTorch(BaseDriftOnline):
+class LSDDDriftOnlineTorch(BaseMultiDriftOnline):
     def __init__(
             self,
             x_ref: Union[np.ndarray, list],
@@ -32,7 +32,8 @@ class LSDDDriftOnlineTorch(BaseDriftOnline):
         x_ref
             Data used as reference distribution.
         ert
-            The expected run-time (ERT) in the absence of drift.
+            The expected run-time (ERT) in the absence of drift. For the multivariate detectors, the ERT is defined
+            as the expected run-time from t=0.
         window_size
             The size of the sliding test-window used to compute the test-statistic.
             Smaller windows focus on responding quickly to severe drift, larger windows focus on