+++ doc/source/conf.py
@@ -55,13 +55,27 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx_autodoc_typehints",
     "sphinxcontrib.apidoc",  # automatically generate API docs, see https://github.com/rtfd/readthedocs.org/issues/1139
+    "sphinxcontrib.bibtex",
     "nbsphinx",
-    "nbsphinx_link",  # for linking notebooks from outside sphinx source root
     "myst_parser",
 ]
 
-# nbsphinx settings
+# -- nbsphinx settings -------------------------------------------------------
 nbsphinx_execute = "auto"
+
+# Create symlinks for example notebooks
+import glob
+nb_files = [os.path.basename(f) for f in glob.glob(os.path.join('examples','*.ipynb')) 
+        if not os.path.basename(f).startswith('temp_')]
+for nb_file in nb_files:
+    target = os.path.join('../../examples', nb_file)
+    if os.path.exists(target):
+        os.remove(target)
+    os.symlink(os.path.join('../doc/source/examples', nb_file), target)
+
+# -- Bibliography ------------------------------------------------------------
+bibtex_bibfiles = ['refs.bib']
+bibtex_default_style = 'unsrtalpha'
 
 # apidoc settings
 apidoc_module_dir = "../../alibi_detect"
@@ -95,6 +109,7 @@ autodoc_mock_imports = [
     "dill",
     "pydantic",
     "catalogue"
+    "numba"
 ]
 
 # Napoleon settings
@@ -249,16 +264,7 @@ intersphinx_mapping = {"https://docs.python.org/": None}
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = True
 
-# from https://github.com/vidartf/nbsphinx-link/blob/master/docs/source/conf.py
-
-# Ensure env.metadata[env.docname]['nbsphinx-link-target']
-# points relative to repo root:
-import os
-
-here = os.path.dirname(__file__)
-repo = os.path.join(here, "..", "..")
-nbsphinx_link_target_root = repo
-
+# -- nbsphinx prolog ---------------------------------------------------------
 # from https://github.com/vidartf/nbsphinx-link/blob/master/docs/source/conf.py for custom tags
 import subprocess
 
@@ -274,11 +280,7 @@ if git_rev:
 
 nbsphinx_prolog = (
     r"""
-{% if env.metadata[env.docname]['nbsphinx-link-target'] %}
-{% set docpath = env.metadata[env.docname]['nbsphinx-link-target'] %}
-{% else %}
-{% set docpath = env.doc2path(env.docname, base='doc/source/') %}
-{% endif %}
+{% set docname = env.doc2path(env.docname, base=False) %}
 
 .. only:: html
 
@@ -286,12 +288,13 @@ nbsphinx_prolog = (
         :format: html
     
     .. nbinfo::
-        This page was generated from `{{ docpath }}`__.
+        This page was generated from `{{ docname }}`__.
     
     __ https://github.com/SeldonIO/alibi-detect/blob/
         """
     + git_rev
-    + r"{{ docpath }}"
+    + "doc/source/"
+    + r"{{ docname }}"
 )
 
 # -- Override order of preference for image formats --------------------------